it can steal Chromium-based browsers pass
This project is designed to extract saved **passwords** from the following browsers:
- **Chrome**
- **Edge**
- **Brave**
- **Opera**

Additionally, it retrieves **cookies** from these browsers:
- **Chrome**
- **Edge**
- **Brave**
- **Opera**
- **Firefox**

This is an open-source project, and you're welcome to modify it as needed. If you create an enhanced version or make any improvements, feel free to share them with me at [amirazazil100@gmail.com]. I'd love to see your contributions!